Fix broken cookies
While working on integrating the latest client into our docs,
I noticed that Amplitude client is not able to create its cookie.
The issue is that in v5.11.0 they introduced the sameSiteCookie
property with default value None. However, it works only wih the
combination of setting Secure flag on a cookie in recent Chrome versions,
which is not done.

This change bumps the version, and fixes the behaviour setting
SameSite attribute to Lax. It matches the latest Amplitude lib
beehaviour (we don't update to the latest because of some breaking changes).

We also pin the client version to avoid such issues in the future.

Change-type: patch
Signed-off-by: Roman Mazur <roman@balena.io>