Changes on how device_id is treated on init
With this changes the consumeUrlParams dont override the client
device_id with the passed one. Instead it returns the passed one
if exists and the clean queryParams string.
Then the amplitude Config is now allowed to recieve a device_id.
So then the client creation happens, if a device_id is passed on Config,
(which should be the one obtained from consumeUrlParams now),
it initiates the client with that device_id, and if not, it initiates
with no device_id, so it creates one.

Since this changes modifies some of the base logic of the
analytics-client, other components implementation should be
revisited to comply with these changes.

Change-type: Major
Signed-off-by: Ezequiel Boehler ezequiel@balena.io