Changed getQueryString to validate domains
Since we only want to append device id and session id when needed,
that is, when navigating cross domain or cross component, now the
getQueryString function checks if there is a passed destination
hostname and if so it compares its main domain to the main domain of
the actual URL the user is on. When a difference between the two exists,
or there was no destination hostname passed it returns both
d_id and s_id params. Otherwise if these 2 are the same, it returns
an empty string.

Change-type: Minor
Signed-off-by: Ezequiel Boehler ezequiel@balena.io